+++ source/ftrack_api/entity/base.py
@@ -3,6 +3,7 @@
 
 from __future__ import absolute_import
 
+from builtins import str
 import abc
 import collections
 import logging
@@ -13,6 +14,12 @@ import ftrack_api.inspection
 import ftrack_api.exception
 import ftrack_api.operation
 from ftrack_api.logging import LazyLogMessage as L
+from future.utils import with_metaclass
+
+
+class _EntityBase(object):
+    '''Base class to allow for mixins, we need a common base.'''
+    pass
 
 
 class DynamicEntityTypeMetaclass(abc.ABCMeta):
@@ -28,10 +35,8 @@ class DynamicEntityTypeMetaclass(abc.ABCMeta):
         return '<dynamic ftrack class \'{0}\'>'.format(self.__name__)
 
 
-class Entity(collections.MutableMapping):
+class Entity(with_metaclass(DynamicEntityTypeMetaclass, _EntityBase, collections.MutableMapping)):
     '''Base class for all entities.'''
-
-    __metaclass__ = DynamicEntityTypeMetaclass
 
     entity_type = 'Entity'
     attributes = None
@@ -93,7 +98,7 @@ class Entity(collections.MutableMapping):
 
 
             # Data represents locally set values.
-            for key, value in data.items():
+            for key, value in list(data.items()):
                 if key in self._ignore_data_keys:
                     continue
 
@@ -144,7 +149,7 @@ class Entity(collections.MutableMapping):
                 )
             )
 
-        for attribute, value in relational_attributes.items():
+        for attribute, value in list(relational_attributes.items()):
             # Finally we set values for "relational" attributes, we need
             # to do this at the end in order to get the create operations
             # in the correct order as the newly created attributes might
@@ -157,7 +162,7 @@ class Entity(collections.MutableMapping):
     def _reconstruct(self, data):
         '''Reconstruct from *data*.'''
         # Data represents remote values.
-        for key, value in data.items():
+        for key, value in list(data.items()):
             if key in self._ignore_data_keys:
                 continue
 
@@ -182,7 +187,7 @@ class Entity(collections.MutableMapping):
         with self.session.auto_populating(False):
             primary_key = ['Unknown']
             try:
-                primary_key = ftrack_api.inspection.primary_key(self).values()
+                primary_key = list(ftrack_api.inspection.primary_key(self).values())
             except KeyError:
                 pass
 
@@ -253,7 +258,7 @@ class Entity(collections.MutableMapping):
         if self.session.auto_populate:
             self._populate_unset_scalar_attributes()
 
-        return super(Entity, self).values()
+        return list(super(Entity, self).values())
 
     def items(self):
         '''Return list of tuples of (key, value) pairs.
@@ -267,7 +272,7 @@ class Entity(collections.MutableMapping):
         if self.session.auto_populate:
             self._populate_unset_scalar_attributes()
 
-        return super(Entity, self).items()
+        return list(super(Entity, self).items())
 
     def clear(self):
         '''Reset all locally modified attribute values.'''