+++ source/ftrack_api/event/expression.py
@@ -1,6 +1,9 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014 ftrack
 
+from builtins import map
+from six import string_types
+from builtins import object
 from operator import eq, ne, ge, le, gt, lt
 
 from pyparsing import (Group, Word, CaselessKeyword, Forward,
@@ -33,7 +36,7 @@ class Parser(object):
     def _construct_parser(self):
         '''Construct and return parser.'''
         field = Word(alphanums + '_.')
-        operator = oneOf(self._operators.keys())
+        operator = oneOf(list(self._operators.keys()))
         value = Word(alphanums + '-_,./*@+')
         quoted_value = quotedString('quoted_value').setParseAction(removeQuotes)
 
@@ -274,7 +277,7 @@ class Condition(Expression):
 
         if (
             self._operator is eq
-            and isinstance(self._value, basestring)
+            and isinstance(self._value, string_types)
             and self._value[-1] == self._wildcard
         ):
             return self._value[:-1] in value