+++ source/ftrack_api/session.py
@@ -2,7 +2,13 @@
 # :copyright: Copyright (c) 2014 ftrack
 
 from __future__ import absolute_import
+from __future__ import division
 
+from builtins import zip
+from builtins import map
+from builtins import str
+from six import string_types
+from builtins import object
 import json
 import logging
 import collections
@@ -76,7 +82,7 @@ class Session(object):
     def __init__(
         self, server_url=None, api_key=None, api_user=None, auto_populate=True,
         plugin_paths=None, cache=None, cache_key_maker=None,
-        auto_connect_event_hub=None, schema_cache_path=None,
+        auto_connect_event_hub=False, schema_cache_path=None,
         plugin_arguments=None
     ):
         '''Initialise session.
@@ -241,7 +247,7 @@ class Session(object):
         )
 
         self._auto_connect_event_hub_thread = None
-        if auto_connect_event_hub in (None, True):
+        if auto_connect_event_hub:
             # Connect to event hub in background thread so as not to block main
             # session usage waiting for event hub connection.
             self._auto_connect_event_hub_thread = threading.Thread(
@@ -249,12 +255,6 @@ class Session(object):
             )
             self._auto_connect_event_hub_thread.daemon = True
             self._auto_connect_event_hub_thread.start()
-
-        # To help with migration from auto_connect_event_hub default changing
-        # from True to False.
-        self._event_hub._deprecation_warning_auto_connect = (
-            auto_connect_event_hub is None
-        )
 
         # Register to auto-close session on exit.
         atexit.register(WeakMethod(self.close))
@@ -491,33 +491,33 @@ class Session(object):
     @property
     def created(self):
         '''Return list of newly created entities.'''
-        entities = self._local_cache.values()
+        entities = list(self._local_cache.values())
         states = ftrack_api.inspection.states(entities)
 
         return [
-            entity for (entity, state) in itertools.izip(entities, states)
+            entity for (entity, state) in zip(entities, states)
             if state is ftrack_api.symbol.CREATED
         ]
 
     @property
     def modified(self):
         '''Return list of locally modified entities.'''
-        entities = self._local_cache.values()
+        entities = list(self._local_cache.values())
         states = ftrack_api.inspection.states(entities)
 
         return [
-            entity for (entity, state) in itertools.izip(entities, states)
+            entity for (entity, state) in zip(entities, states)
             if state is ftrack_api.symbol.MODIFIED
         ]
 
     @property
     def deleted(self):
         '''Return list of deleted entities.'''
-        entities = self._local_cache.values()
+        entities = list(self._local_cache.values())
         states = ftrack_api.inspection.states(entities)
 
         return [
-            entity for (entity, state) in itertools.izip(entities, states)
+            entity for (entity, state) in zip(entities, states)
             if state is ftrack_api.symbol.DELETED
         ]
 
@@ -628,7 +628,7 @@ class Session(object):
 
         '''
         if not identifying_keys:
-            identifying_keys = data.keys()
+            identifying_keys = list(data.keys())
 
         self.logger.debug(L(
             'Ensuring entity {0!r} with data {1!r} using identifying keys '
@@ -647,7 +647,7 @@ class Session(object):
         for identifying_key in identifying_keys:
             value = data[identifying_key]
 
-            if isinstance(value, basestring):
+            if isinstance(value, string_types):
                 value = '"{0}"'.format(value)
 
             elif isinstance(
@@ -682,7 +682,7 @@ class Session(object):
 
             # Update entity if required.
             updated = False
-            for key, target_value in data.items():
+            for key, target_value in list(data.items()):
                 if entity[key] != target_value:
                     entity[key] = target_value
                     updated = True
@@ -715,7 +715,7 @@ class Session(object):
         self.logger.debug(L('Get {0} with key {1}', entity_type, entity_key))
 
         primary_key_definition = self.types[entity_type].primary_key_attributes
-        if isinstance(entity_key, basestring):
+        if isinstance(entity_key, string_types):
             entity_key = [entity_key]
 
         if len(entity_key) != len(primary_key_definition):
@@ -762,7 +762,7 @@ class Session(object):
         # Check cache for existing entity emulating
         # ftrack_api.inspection.identity result object to pass to key maker.
         cache_key = self.cache_key_maker.key(
-            (str(entity_type), map(str, entity_key))
+            (str(entity_type), list(map(str, entity_key)))
         )
         self.logger.debug(L(
             'Checking cache for entity with key {0}', cache_key
@@ -1065,7 +1065,7 @@ class Session(object):
 
             primary_key_definition = reference_entity.primary_key_attributes
             entity_keys = [
-                ftrack_api.inspection.primary_key(entity).values()
+                list(ftrack_api.inspection.primary_key(entity).values())
                 for entity in entities_to_process
             ]
 
@@ -1129,7 +1129,7 @@ class Session(object):
                     payload = OperationPayload({
                         'action': 'create',
                         'entity_type': operation.entity_type,
-                        'entity_key': operation.entity_key.values(),
+                        'entity_key': list(operation.entity_key.values()),
                         'entity_data': entity_data
                     })
 
@@ -1146,7 +1146,7 @@ class Session(object):
                     payload = OperationPayload({
                         'action': 'update',
                         'entity_type': operation.entity_type,
-                        'entity_key': operation.entity_key.values(),
+                        'entity_key': list(operation.entity_key.values()),
                         'entity_data': entity_data
                     })
 
@@ -1156,7 +1156,7 @@ class Session(object):
                     payload = OperationPayload({
                         'action': 'delete',
                         'entity_type': operation.entity_type,
-                        'entity_key': operation.entity_key.values()
+                        'entity_key': list(operation.entity_key.values())
                     })
 
                 else:
@@ -1206,7 +1206,7 @@ class Session(object):
         updates_map = set()
         for payload in reversed(batch):
             if payload['action'] in ('update', ):
-                for key, value in payload['entity_data'].items():
+                for key, value in list(payload['entity_data'].items()):
                     if key == '__entity_type__':
                         continue
 
@@ -1221,7 +1221,7 @@ class Session(object):
         # Remove NOT_SET values from entity_data.
         for payload in batch:
             entity_data = payload.get('entity_data', {})
-            for key, value in entity_data.items():
+            for key, value in list(entity_data.items()):
                 if value is ftrack_api.symbol.NOT_SET:
                     del entity_data[key]
 
@@ -1230,7 +1230,7 @@ class Session(object):
         for payload in batch:
             entity_data = payload.get('entity_data')
             if entity_data is not None:
-                keys = entity_data.keys()
+                keys = list(entity_data.keys())
                 if not keys or keys == ['__entity_type__']:
                     continue
 
@@ -1272,7 +1272,7 @@ class Session(object):
             # remain as needed for cache retrieval on new entities.
             with self.auto_populating(False):
                 with self.operation_recording(False):
-                    for entity in self._local_cache.values():
+                    for entity in list(self._local_cache.values()):
                         for attribute in entity:
                             if attribute not in entity.primary_key_attributes:
                                 del entity[attribute]
@@ -1292,7 +1292,7 @@ class Session(object):
             # keys on entities that were merged.
             with self.auto_populating(False):
                 with self.operation_recording(False):
-                    for entity in self._local_cache.values():
+                    for entity in list(self._local_cache.values()):
                         entity.clear()
 
     def rollback(self):
@@ -1322,7 +1322,7 @@ class Session(object):
                     ):
                         entity_key = str((
                             str(operation.entity_type),
-                            operation.entity_key.values()
+                            list(operation.entity_key.values())
                         ))
                         try:
                             self.cache.remove(entity_key)
@@ -1330,7 +1330,7 @@ class Session(object):
                             pass
 
                 # Clear locally stored modifications on remaining entities.
-                for entity in self._local_cache.values():
+                for entity in list(self._local_cache.values()):
                     entity.clear()
 
         self.recorded_operations.clear()
@@ -1383,7 +1383,7 @@ class Session(object):
         with open(schema_cache_path, 'r') as schema_file:
             schemas = json.load(schema_file)
             hash_ = hashlib.md5(
-                json.dumps(schemas, sort_keys=True)
+                json.dumps(schemas, sort_keys=True).encode('utf-8')
             ).hexdigest()
 
         return schemas, hash_
@@ -1602,22 +1602,6 @@ class Session(object):
             synchronous=True
         )
 
-    @ftrack_api.logging.deprecation_warning(
-        'Session._call is now available as public method Session.call. The '
-        'private method will be removed in version 2.0.'
-    )
-    def _call(self, data):
-        '''Make request to server with *data* batch describing the actions.
-
-        .. note::
-
-            This private method is now available as public method
-            :meth:`entity_reference`. This alias remains for backwards
-            compatibility, but will be removed in version 2.0.
-
-        '''
-        return self.call(data)
-
     def call(self, data):
         '''Make request to server with *data* batch describing the actions.'''
         url = self._server_url + '/api'
@@ -1709,9 +1693,9 @@ class Session(object):
             }
 
         if isinstance(item, OperationPayload):
-            data = dict(item.items())
+            data = dict(list(item.items()))
             if "entity_data" in data:
-                for key, value in data["entity_data"].items():
+                for key, value in list(data["entity_data"].items()):
                     if isinstance(value, ftrack_api.entity.base.Entity):
                         data["entity_data"][key] = self.entity_reference(value)
 
@@ -1783,26 +1767,6 @@ class Session(object):
 
         return reference
 
-    @ftrack_api.logging.deprecation_warning(
-        'Session._entity_reference is now available as public method '
-        'Session.entity_reference. The private method will be removed '
-        'in version 2.0.'
-    )
-    def _entity_reference(self, entity):
-        '''Return entity reference that uniquely identifies *entity*.
-
-        Return a mapping containing the __entity_type__ of the entity along
-        with the key, value pairs that make up it's primary key.
-
-        .. note::
-
-            This private method is now available as public method
-            :meth:`entity_reference`. This alias remains for backwards
-            compatibility, but will be removed in version 2.0.
-
-        '''
-        return self.entity_reference(entity)
-
     def decode(self, string):
         '''Return decoded JSON *string* as Python object.'''
         with self.operation_recording(False):
@@ -1834,10 +1798,7 @@ class Session(object):
 
         # Filter.
         if filter_inaccessible:
-            locations = filter(
-                lambda location: location.accessor,
-                locations
-            )
+            locations = [location for location in locations if location.accessor]
 
         # Sort by priority.
         locations = sorted(
@@ -2077,7 +2038,7 @@ class Session(object):
         container_components = []
 
         for component in components:
-            if 'members' in component.keys():
+            if 'members' in list(component.keys()):
                 container_components.append(component)
             else:
                 standard_components.append(component)
@@ -2102,7 +2063,7 @@ class Session(object):
             availability = base_availability.copy()
             availabilities.append(availability)
 
-            is_container = 'members' in component.keys()
+            is_container = 'members' in list(component.keys())
             if is_container and len(component['members']):
                 member_availabilities = self.get_component_availabilities(
                     component['members'], locations=locations
@@ -2111,7 +2072,7 @@ class Session(object):
                 for member, member_availability in zip(
                     component['members'], member_availabilities
                 ):
-                    for location_id, ratio in member_availability.items():
+                    for location_id, ratio in list(member_availability.items()):
                         availability[location_id] += (
                             ratio * multiplier
                         )
@@ -2121,7 +2082,7 @@ class Session(object):
                     if location_id in availability:
                         availability[location_id] = 100.0
 
-            for location_id, percentage in availability.items():
+            for location_id, percentage in list(availability.items()):
                 # Avoid quantization error by rounding percentage and clamping
                 # to range 0-100.
                 adjusted_percentage = round(percentage, 9)
@@ -2129,36 +2090,6 @@ class Session(object):
                 availability[location_id] = adjusted_percentage
 
         return availabilities
-
-    @ftrack_api.logging.deprecation_warning(
-        'Session.delayed_job has been deprecated in favour of session.call. '
-        'Please refer to the release notes for more information.'
-    )
-    def delayed_job(self, job_type):
-        '''Execute a delayed job on the server, a `ftrack.entity.job.Job` is returned.
-
-        *job_type* should be one of the allowed job types. There is currently
-        only one remote job type "SYNC_USERS_LDAP".
-        '''
-        if job_type not in (ftrack_api.symbol.JOB_SYNC_USERS_LDAP, ):
-            raise ValueError(
-                u'Invalid Job type: {0}.'.format(job_type)
-            )
-
-        operation = {
-            'action': 'delayed_job',
-            'job_type': job_type.name
-        }
-
-        try:
-            result = self.call(
-                [operation]
-            )[0]
-
-        except ftrack_api.exception.ServerError as error:
-            raise
-
-        return result['data']
 
     def get_widget_url(self, name, entity=None, theme=None):
         '''Return an authenticated URL for widget with *name* and given options.
@@ -2184,7 +2115,7 @@ class Session(object):
         if entity:
             operation['entity_type'] = entity.entity_type
             operation['entity_key'] = (
-                ftrack_api.inspection.primary_key(entity).values()
+                list(ftrack_api.inspection.primary_key(entity).values())
             )
 
         try:
@@ -2247,7 +2178,7 @@ class Session(object):
         is a FileComponent, and deleted if it is a file path. You can specify
         True or False to change this behavior.
         '''
-        if isinstance(media, basestring):
+        if isinstance(media, string_types):
             # Media is a path to a file.
             server_location = self.get(
                 'Location', ftrack_api.symbol.SERVER_LOCATION_ID