+++ test/unit/event/test_hub.py
@@ -9,6 +9,7 @@ import subprocess
 import sys
 
 import pytest
+from flaky import flaky
 
 import ftrack_api.event.hub
 import ftrack_api.event.subscriber
@@ -223,10 +224,9 @@ def test_connect_missing_required_transport(session, mocker, caplog):
 
     with pytest.raises(ftrack_api.exception.EventHubConnectionError):
         event_hub.connect()
-
-    logs = caplog.records()
+    
     assert (
-        'Server does not support websocket sessions.' in str(logs[-1].exc_info)
+        'Server does not support websocket sessions.' in str(caplog.text())
     )
 
 
@@ -443,6 +443,7 @@ def test_unsubscribe(event_hub):
     assert_callbacks(event_hub, [callback_b])
 
 
+@flaky(max_runs=2, min_passes=1)
 def test_unsubscribe_whilst_disconnected(event_hub):
     '''Unsubscribe whilst disconnected.'''
     identifier = event_hub.subscribe('topic=test', None)
@@ -558,7 +559,14 @@ def test_publish_logs_other_errors(event_hub, caplog, mocker):
     event_hub.publish(event)
 
     expected = 'Error sending event {0}.'.format(event)
-    messages = [record.getMessage().strip() for record in caplog.records()]
+
+
+    if not isinstance(caplog.records, list):
+        records = caplog.records()
+    else:
+        records = caplog.records
+
+    messages = [record.getMessage().strip() for record in records]
     assert expected in messages, 'Expected log message missing in output.'